+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Grammar updates for C3 0.7.4.
 
 - Added `enum Foo : const` 
 - Update doc contract grammar
+- Update real and integer literals
 
 ## 0.8.1
 Grammar updates for C3 0.7.3.